+++ src/FormContext.ts
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import { ValidateMessages } from './interface';
+
+interface FormContextProps {
+  validateMessages?: ValidateMessages;
+}
+
+const Context = React.createContext<FormContextProps>({});
